Fix bug in the inclusion of FrancisOpts
